Fix issue with readonly fields for is_sox_restricted = False
Move helping methods to separate utils module
