+++ src/ggrc/models/mixins/with_custom_restrictions.py
@@ -4,13 +4,11 @@
 """Contains the WithCustomRestrictions mixin"""
 import logging
 
-from datetime import datetime
-from datetime import date
-
 from ggrc.models import reflection
 from ggrc.models.mixins.statusable import Statusable
 from ggrc.rbac import permissions
 from ggrc.utils import benchmark
+from ggrc.utils import json_comparator
 
 logger = logging.getLogger(__name__)
 
@@ -81,7 +79,7 @@ class WithCustomRestrictions(object):
     # type: () -> tuple
     """Get list of all readonly fields for object {self} for user current
     user"""
-    if not self._is_sox_restricted:
+    if not self._is_sox_restricted():
       return tuple()
     for field_name, restrictions_dict in self._restriction_condition.items():
       obj_state = getattr(self, field_name)
@@ -91,58 +89,6 @@ class WithCustomRestrictions(object):
           return read_only_fields
     return tuple()
 
-  def _are_lists_equal(self, list1, list2):
-    """Compare two lists and return a result: True if lists do not have
-    updated values."""
-    if len(list1) != len(list2):
-      return False
-    if not list1:
-      return True
-    if isinstance(list1[0], dict):
-      # we need it to have sorted values that we able to compare
-      dict1 = {item.get('id'): item for item in list1}
-      dict2 = {item.get('id'): item for item in list2}
-      return self._are_dicts_equal(dict1, dict2)
-    return sorted(list1) == sorted(list2)
-
-  def _are_dicts_equal(self, dict1, dict2):
-    """Compare two dictionaries and return a result: True if dictionaries do
-    not have different key-value pair. Comparison does not count new key-value
-    pairs that does not contain one of dictionaries."""
-    if len(dict1.keys()) > len(dict2.keys()):
-      dict1, dict2 = dict2, dict1
-    for field, value in dict1.items():
-      if str(field).startswith("_"):
-        # Debug information does not related to model
-        continue
-      if field in dict2:
-        if not self._are_fields_equal(dict2[field], value):
-          return False
-    return True
-
-  def _are_fields_equal(self, obj_field, src_field):
-    """Compare tho objects and their content. If an object is a type of
-     List of Dict, then each item of it will be compared. Returns True if
-     objects are equal to each other."""
-    if isinstance(obj_field, dict):
-      return self._are_dicts_equal(obj_field, src_field)
-    elif isinstance(obj_field, list):
-      return self._are_lists_equal(obj_field, src_field)
-
-    obj_field = self.convert_to_string(obj_field)
-    src_field = self.convert_to_string(src_field)
-
-    return obj_field == src_field
-
-  @staticmethod
-  def convert_to_string(custom_value):
-    """Convert custom types to string to simplify comparison"""
-    if isinstance(custom_value, datetime):
-      return custom_value.strftime("%Y-%m-%dT%H:%M:%S")
-    if isinstance(custom_value, date):
-      return custom_value.strftime("%Y-%m-%d")
-    return custom_value
-
   def is_updating_readonly_fields(self, src):
     """Check is {src} going to update fields that is readonly for current
     {user}"""
@@ -150,8 +96,8 @@ class WithCustomRestrictions(object):
                        if field not in self.mapping_restrictions()])
     json_obj = self.log_json()
     for field in ro_fields:
-      if not self._are_fields_equal(json_obj.get(field, None),
-                                    src.get(field, None)):
+      if not json_comparator.fields_equal(json_obj.get(field, None),
+                                          src.get(field, None)):
         return True
     return False
