+++ test/integration/ggrc/models/mixins/test_with_custom_restrictions.py
@@ -71,20 +71,11 @@ class TestWithCustomRestrictions(TestCase):
         object_type="Assessment",
     ).first()
 
-    program = factories.ProgramFactory()
-
-    audit = factories.AuditFactory(program=program)
-    factories.RelationshipFactory(
-        source=audit,
-        destination=program,
-    )
+    audit = factories.AuditFactory()
     self.assign_person(audit, acr_audit, user.id)
 
     assessment = factories.AssessmentFactory(audit=audit, sox_302_enabled=True)
     self.assign_person(assessment, acr_assmnt, user.id)
-    factories.RelationshipFactory(
-        source=audit, destination=assessment
-    )
 
     self.assertFalse(assessment.is_user_role_restricted(user))
